+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.19.1] 2022-03-08
+
+### Fixed
+
+- filter section for tags
+- published sort order
+
 ## [2.19.0] 2022-03-03
 
 ### Changed